add incremental-cargo-update
This is a major part of the tooling required to implement b/244594644.
The only remaining required bit is a script which, on a regular basis:
- syncs a ChromeOS tree
- runs this
- uploads the CL that running this produced, if any changes were made

These bits should be easy to do, though will be in an internal team
repo. Might as well FOSS-ify the bits that can easily be FOSS-ified...

BUG=b:244594644
TEST=Ran on rust_crates.

Change-Id: Ic2dbd40f7e228e5c2d3bd318d4ca6d01107cf052
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/rust_crates/+/4136832
Commit-Queue: George Burgess <gbiv@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>
Tested-by: George Burgess <gbiv@chromium.org>